Ceci s'appelle la sécurité

+++ src/main/java/fr/epsi/springHalloween/springHalloween/repository/ServiceRepository.java
@@ -3,5 +3,5 @@ package fr.epsi.springHalloween.springHalloween.repository;
 import fr.epsi.springHalloween.springHalloween.entity.Service;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ServiceRepository extends CrudRepository<Service, Long> {
+public interface ServiceRepository extends CrudRepository<Service, Integer> {
 }
